Remove clock dependency
When exploring the new version of `Data.Pool` I noticed it was getting
monotonic time from `GHC.Clock`. I did not know this module exists, but
it is a very simple foreign function interface. There is no reason to
rely on the `clock` package when this exists.

This also adds a changelog to the repo.